% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{The Dijkstras algortithm takes a graph and an initial node and calculates the shortest path from the initial node to every other node.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{a dataframe.}

\item{init_node}{a number.}
}
\value{
The shortest path to every other node from the starting node as a vector.
}
\description{
The Dijkstras algortithm takes a graph and an initial node and calculates the shortest path from the initial node to every other node.
}
\examples{
dijkstra()
dijkstra()
}
