% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm to find the greatest common divisor of two numbers.}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The greatest common divisor of both inputs which is a number.
}
\description{
Euclidean algorithm to find the greatest common divisor of two numbers.
}
\section{Reference}{
 https://en.wikipedia.org/wiki/Euclidean_algorithm
}

\examples{
euclidean(252, 105)
euclidean(100, 1000)
}
